fix: fixed operations with posts

+++ backend/app/posts/post.controller.js
@@ -5,32 +5,35 @@ const authorize = require('../_middleware/authorize');
 const Role = require('../_helpers/role');
 
 // routes
-router.post('/', authorize(), create);
+router.post('/', create);
 router.get('/', getAll);
-router.put('/:id', authorize(), update);
-router.delete('/:id', authorize(), _delete);
+router.put('/:id', update);
+router.delete('/:id', _delete);
 
 module.exports = router;
 
 function create(req, res, next) {
+  authorize(),
   postService.create(req.body)
-    .then(post => res.json({post: post}))
+    .then(post => res.json(post))
     .catch(err => next(err));
 }
 
 function getAll(req, res, next) {
   postService.getAll()
-    .then(posts => res.json({posts: posts}))
+    .then(posts => res.json(posts))
     .catch(err => next(err));
 }
 
 function update(req, res, next) {
+  authorize(),
   postService.update(req.params.id, req.body)
-    .then(post => res.json({post: post}))
+    .then(post => res.json(post))
     .catch(err => next(err));
 }
 
 function _delete(req, res, next) {
+  authorize(),
   postService.delete(req.params.id)
     .then(() => res.json({}))
     .catch(err => next(err));
