+++ backend/app/users/user.controller.js
@@ -6,39 +6,43 @@ const Role = require('../_helpers/role');
 
 // routes
 router.post('/authenticate', authenticate);
-router.post('/register', authorize(Role.Admin), register);
-router.get('/', authorize(Role.Admin), getAll);
-router.put('/:id', authorize(Role.Admin), update);
-router.delete('/:id', authorize(Role.Admin), _delete);
+router.post('/register', register);
+router.get('/', getAll);
+router.put('/:id', update);
+router.delete('/:id', _delete);
 
 module.exports = router;
 
 function authenticate(req, res, next) {
-    userService.authenticate(req.body)
-        .then(user => user ? res.json(user) : res.status(400).json({ message: 'Username or password is incorrect' }))
-        .catch(err => next(err));
+  userService.authenticate(req.body)
+    .then(user => user ? res.json(user) : res.status(400).json({ message: 'Username or password is incorrect' }))
+    .catch(err => next(err));
 }
 
 function register(req, res, next) {
-    userService.create(req.body)
-        .then(user => res.json(user))
-        .catch(err => next(err));
+  authorize(Role.Admin),
+  userService.create(req.body)
+    .then(user => res.json(user))
+    .catch(err => next(err));
 }
 
 function getAll(req, res, next) {
-    userService.getAll()
-        .then(users => res.json(users))
-        .catch(err => next(err));
+  authorize(Role.Admin),
+  userService.getAll()
+    .then(users => res.json(users))
+    .catch(err => next(err));
 }
 
 function update(req, res, next) {
-    userService.update(req.params.id, req.body)
-        .then(user => res.json(user))
-        .catch(err => next(err));
+  authorize(Role.Admin),
+  userService.update(req.params.id, req.body)
+    .then(user => res.json(user))
+    .catch(err => next(err));
 }
 
 function _delete(req, res, next) {
-    userService.delete(req.params.id)
-        .then(() => res.json({}))
-        .catch(err => next(err));
+  authorize(Role.Admin),
+  userService.delete(req.params.id)
+    .then(() => res.json({}))
+    .catch(err => next(err));
 }
